+++ test/index.ts
@@ -2849,6 +2849,16 @@ describe('std', () => {
 				ARR([STR('c'), NUM(3)])
 			]));
 		});
+
+		test.concurrent('merge', async () => {
+			const res = await exe(`
+			let o1 = { a: 1; b: 2; }
+			let o2 = { b: 3; c: 4; }
+
+			<: Obj:merge(o1, o2)
+			`);
+			eq(res, utils.jsToVal({ a: 1, b: 3, c: 4}));
+		});
 	});
 
 	describe('Str', () => {
@@ -2876,6 +2886,17 @@ describe('std', () => {
 			<: "".codepoint_at(0)
 			`);
 			eq(res, NULL);
+		});
+	});
+
+	describe('Error', () => {
+		test.concurrent('create', async () => {
+			eq(
+				await exe(`
+				<: Error:create('ai', {chan: 'kawaii'})
+				`),
+				ERROR('ai', OBJ(new Map([['chan', STR('kawaii')]])))
+			);
 		});
 	});